feat(schemaEdit): major improvements to schema Editor, aslo allows editing schemas to be finalised
