+++ apps/Conduit-UI/src/pages/database/introspection/[id].tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import SchemaEditor from '../../../components/database/schemas/SchemaEditor';
+
+const IntrospectionSchemaEditor = () => {
+  return <SchemaEditor introspection={true} />;
+};
+
+export default IntrospectionSchemaEditor;
